# Refactor: Optimize test performance for faster CI execution